refactor: food-service 프로젝트 전체파일 app 모듈 파일로 이동
