+++ app/src/main/asciidoc/content/2_auth.adoc
