+++ app/src/main/java/com/food/AppApplication.java
@@ -1,4 +1,4 @@
-package com.food.app;
+package com.food;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
