+++ app/src/main/java/com/food/scanner/DefaultRequestPathScanner.java
