+++ app/src/test/java/com/food/mock/store/MockStoreImage.java
